chore(rename): Renamed the context directory to chores
