+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import { ContextProvider } from "./context/Context"
+import { ContextProvider } from "./contexts/Context"
 import App from './App.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
